Refactored and cleaned up file structure, removing unneeded files
